persistencia casi lista
